pyproject.toml
[tool.poetry]
name = "forum"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
markdown = "^3.6"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"
pymongo = "^4.8.0"
mongoengine = "^0.28.2"
celery = {extras = ["redis"], version = "^5.4.0"}

[tool.poetry.group.dev.dependencies]
pylint = "^3.2.5"
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"